Convert nets per capita to annual nets per capital delivered

Convert nets per capita to annual nets per capital delivered

Usage

crop_to_distribution(
  crop,
  distribution_freq,
  half_life,
  net_loss_function = net_loss_map,
  ...
)

Arguments

crop

Nets per capita

distribution_freq

A single distribution frequency of nets in days. Default = 3*365 (3 years).

half_life

Net loss half life (days)

net_loss_function

Option to choose between exponential net loss (net_loss_exp) or the smooth-compact net loss function from Bertozzi-Villa, Amelia, et al. Nature communications 12.1 (2021): 1-12 (net_loss_map). Default = net_loss_map.

...

additional arguments for the net_loss_function

Value

Annual nets per capita delivered