Estimate the malaraisimulation input distribution from usage assumes an exponentially distributed net loss function and randomly correlated net distribution.

Estimate the malaraisimulation input distribution from usage assumes an exponentially distributed net loss function and randomly correlated net distribution.

Usage

usage_to_model_distribution(
  usage,
  usage_timesteps,
  distribution_timesteps,
  distribution_lower = rep(0, length(distribution_timesteps)),
  distribution_upper = rep(1, length(distribution_timesteps)),
  mean_retention = 365 * 5
)

Arguments

usage

Target usage

usage_timesteps

Target usage time points

distribution_timesteps

A vector of distribution time steps

distribution_lower

Lower bound on distributions (default = 0)

distribution_upper

Upper bound on distribution (default = 1)

mean_retention

The average duration of net retention (days)